SCRFD: Moving anchors to utils.
