+++ depthai_nodes/node/parsers/utils/scrfd.py
@@ -1,6 +1,39 @@
+from typing import Dict, List, Tuple
+
 import numpy as np
 
 from depthai_nodes.node.parsers.utils.nms import nms
+
+
+def compute_anchor_centers(
+    strides: List[int], input_size: Tuple[int, int], num_anchors: int
+) -> Dict[int, np.ndarray]:
+    """Compute the anchor centers for a given list of strides, input size, and number of
+    anchors.
+
+    @param strides: List of strides.
+    @type strides: List[int]
+    @param input_size: Input size.
+    @type input_size: Tuple[int, int]
+    @param num_anchors: Number of anchors.
+    @type num_anchors: int
+    @return: Dictionary of anchor centers.
+    @rtype: Dict[int, np.ndarray]
+    """
+    anchor_centers_dict = {}
+    for stride in strides:
+        height = input_size[0] // stride
+        width = input_size[1] // stride
+        anchor_centers = np.stack(np.mgrid[:height, :width][::-1], axis=-1).astype(
+            np.float32
+        )
+        anchor_centers = (anchor_centers * stride).reshape((-1, 2))
+        if num_anchors > 1:
+            anchor_centers = np.stack([anchor_centers] * num_anchors, axis=1).reshape(
+                (-1, 2)
+            )
+        anchor_centers_dict[stride] = anchor_centers
+    return anchor_centers_dict
 
 
 def distance2bbox(points, distance, max_shape=None):
